Mummi and bilbo contribution!! Curitiba contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, sjpablo (Pawel Kulesza), fuel (Benjamin Kuhl), r00tkiller (Alexandre),  vapaamies (Vladislav Javadov), epryka (Vova - Vladimir), theaifam5 (Mateus), doraemon (currently... also gosu - hiiamgosu,  dum - noticemesenpaiiamdum..., litAsHellhahah, icarryyourcurry), phanxgames (Henry Price), mushmouth (Patrick McFarling),  sonnystar (sonnystaminecraft, Vlad Ionescu), drmjg, ThomasBS (Thomas Borregaard Sørensen - homasbs.com, KM - Kompetence Match: kmatch.dk), pavankataria (Pavan Kataria),  0tho, 7ru57, abbyyyyy, afterdesign, alaiqka, aleksy, alexcode, andrevdb, anita, austingirl, ayvatan, bazz3l, b3n0it, beee, biaz, bigboom, blackburn1911, blackhatisme, bmw, boza, breitembach, camwhite (Cameron White - YOSHICREW | ng2/node + mixing fire, Yoshicrew.com-out-loud contribution...), cary, caste, chrifrachid, code4life, codeotea, cookiesliyr, chubz, cubor, danielv, daratho, david_joffe, denvercowboy, dgazail, diagram, dimer, dopedopedope, dorumd, dzetko, dzhekson6000, egarcia, elina, eugene_esca, euphemisminc, fauker, ffabiorj, flowdeep, flythiago, fork, frankly444, georgeb, gillie, gzj71ck4elm2, heuy_freeman, hailviral1, harekrish, hudwan, igsky, intelijens (Michael Bui), irondragon000, ishaan007lucky, jamesxism, jarzi, jaycee, jenny0030, jeraqiv, jessy, jilizart, johnyblaze, jookeer, jrtruji, kaeris, karrmarr, kaznowski, keepingtrck, kimliksiz, kissloryshy, koen, kozmo, kp0c, kranalex, LaffayetteBrittoPhoto (streamer with its C300 photo), lagroid, learningcode, liveedu_support, livecodingtv..., lukey_reborn78, lyndroid, magrib, marcie, marianyc, matroskinb, michaeljcalkins, michaelphp, milodewit, milurabbit, mintisbad, monicaa, nickolans, nuke, paezrice, peter_cados, patrick1, pschoeller, pudottapommin, purpleizlegitx, rex2, rooney2, rozumnyib, sam-anarch, samuraijackdavidson, sanyakozmey, Shmellyorc, silvaesouza, simplyronan, siox, skycis, skyfoxx, smart6307 (thunder will/canal2048), sova, specter, systemis, szy, tarikogut, teras, thiagovgamg,  tonycode, trello, usandy, userjoao, vkaracic, v0n3x, wanket, wapaca, web2ls, xalbia, yanaseattle, yeoupooh, zaaz, zakaria98, zann, zarytskiy, zeljko16 and zzzel (Marko Zel)  contribution!! jiangyubo530, kozmo, contribution!! Twitch contribution!! KindaFunnyGames (Greg, Colin, Tim, Nick and Kevin contribution..), AlfredoPlays (and his mates and Fwends playing Siege: Above_Perfection, koopz, Macky43 ... ), SoulFoam, Tectuma (GothicatCat1(Stacie), Tectuma(Kevin), Matty and Dylan), 2016pocket, 4737, Bialyorzeu, Bdjqpee, Bonejacker, cary205, cpmgamingltd, Curseappbot, Daratho, Dat_nerd82, Deathknightruin, Dulocka, esmc103, Feather1231, GameDevCompany (Björn Swedish-Speaking-Finlandian), Gmthiago, GrubbyJuicde, hailviral1,  Handmadehero (Casey Muratori), HardlyDifficult, jhovgaard, Juanpablothefifth, Kennedys3, kimliksiz, Letorusii, Matriz_, Mattyplays6, michaeljcalkins, MOZG888VADIM, Niabi226, nothings2 (Sean Barrett), Ojunbamo, pixel_maniacs, Prafax, Pukateiubeste, Scutalu228, Sanyadoom, Shathan, Shmellyorc, Sinnedk, Skintrader_pl, Snipe0803, sundz98 (from Sweden), Sweegzv, talus7, Trump88, Vjzelynxykat, Wheeze202 (and hosting grumbul (British accent?)), XanderGamingTwich (Xander Jenkins) and Yarpenzid contribution!! Beam.pro contribution!! DasKontinuum (Thanks for the Gesundheit Greeting while doing the Plain Edit/Edit File Great Refactoring), h1dd3n, RoxenRox and Xellar contribution!! bisba contribution!! Carol, Lúcio and Lucinho contribution!! Santos contribution!!  Paulo Picciani contribution!! Upwork contribution!!
bilbo at veriverse
i7 6700 32GB RAM Windows 10 Pro
(booting also macOS Sierra, Hyper-V and Ubuntu 16.04 Unity Numix Theme Numix Circle Icons)

+++ appseed/aura/primitive/primitive_var.cpp
@@ -3977,7 +3977,7 @@ end:
    }
    else
    {
-      operator = (atoi(string(pszStart, pszParse - pszStart)));
+      operator = (atoll(string(pszStart, pszParse - pszStart)));
    }
    psz = pszParse;
 }
